+++ linera-views/src/key_value_store_view.rs
@@ -25,6 +25,7 @@ pub struct KeyValueStoreView<C> {
     context: C,
     was_cleared: bool,
     updates: BTreeMap<Vec<u8>, Option<Vec<u8>>>,
+    stored_hash: Option<HashOutput>,
     hash: Option<HashOutput>,
 }
 
@@ -45,6 +46,7 @@ where
             context,
             was_cleared: false,
             updates: BTreeMap::new(),
+            stored_hash: hash,
             hash,
         })
     }
@@ -52,7 +54,7 @@ where
     fn rollback(&mut self) {
         self.was_cleared = false;
         self.updates.clear();
-        self.hash = None;
+        self.hash = self.stored_hash;
     }
 
     fn flush(&mut self, batch: &mut Batch) -> Result<(), ViewError> {
@@ -74,10 +76,13 @@ where
                 }
             }
         }
-        let key = self.context.base_tag(KeyTag::Hash as u8);
-        match self.hash {
-            None => batch.delete_key(key),
-            Some(hash) => batch.put_key_value(key, &hash)?,
+        if self.stored_hash != self.hash {
+            let key = self.context.base_tag(KeyTag::Hash as u8);
+            match self.hash {
+                None => batch.delete_key(key),
+                Some(hash) => batch.put_key_value(key, &hash)?,
+            }
+            self.stored_hash = self.hash;
         }
         Ok(())
     }
@@ -103,6 +108,7 @@ where
             context,
             was_cleared: false,
             updates: BTreeMap::new(),
+            stored_hash: None,
             hash: None,
         }
     }
